refactor: implement BeanScanner
