+++ nextstep-di/src/main/java/nextstep/di/scanner/BeanScanner.java
@@ -1,0 +1,5 @@
+package nextstep.di.scanner;
+
+public interface BeanScanner {
+    void scanBeans(Object... basePackage);
+}
