+++ nextstep-di/src/test/java/nextstep/di/factory/BeanFactoryTest.java
@@ -1,13 +1,12 @@
 package nextstep.di.factory;
 
-import java.util.Set;
-
 import com.google.common.collect.Sets;
 import nextstep.di.factory.circularreference.CircularReferenceController;
 import nextstep.di.factory.example.MyQnaService;
 import nextstep.di.factory.example.NoDefaultConstructorController;
 import nextstep.di.factory.example.QnaController;
 import nextstep.di.factory.notbean.InjectNotBean;
+import nextstep.di.scanner.ClasspathBeanScanner;
 import nextstep.exception.BeanCreateFailException;
 import nextstep.exception.CircularReferenceException;
 import nextstep.exception.DefaultConstructorFindFailException;
