+++ nextstep-di/src/test/java/nextstep/di/scanner/ClasspathBeanScannerTest.java
@@ -1,27 +1,31 @@
-package nextstep.di.factory;
+package nextstep.di.scanner;
 
 
 import java.util.Arrays;
 import java.util.HashSet;
 import java.util.Set;
 
+import nextstep.di.factory.BeanFactory;
 import nextstep.di.factory.example.JdbcQuestionRepository;
 import nextstep.di.factory.example.JdbcUserRepository;
 import nextstep.di.factory.example.MyQnaService;
 import nextstep.di.factory.example.QnaController;
 import org.junit.jupiter.api.Test;
 
-import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
 
 class ClasspathBeanScannerTest {
     @Test
     void getPreInstantiateClass() {
-        ClasspathBeanScanner classpathBeanScanner = new ClasspathBeanScanner(new BeanFactory());
+        BeanFactory beanFactory = new BeanFactory();
+        ClasspathBeanScanner classpathBeanScanner = new ClasspathBeanScanner(beanFactory);
 
         Set<Class<?>> expected = new HashSet<>(Arrays.asList(QnaController.class, MyQnaService.class,
                 JdbcQuestionRepository.class, JdbcUserRepository.class));
-        Set<Class<?>> actual = classpathBeanScanner.scanBeans("nextstep.di.factory.example");
+        classpathBeanScanner.scanBeans("nextstep.di.factory.example");
+        beanFactory.initialize();
 
-        assertThat(expected).isEqualTo(actual);
+        expected.stream()
+                .forEach(clazz -> assertNotNull(beanFactory.getBean(clazz)));
     }
 }
