+++ nextstep-mvc/src/main/java/nextstep/mvc/tobe/AnnotationHandlerMapping.java
@@ -12,8 +12,6 @@ import com.google.common.collect.Maps;
 import com.google.common.collect.Sets;
 import nextstep.ApplicationContext;
 import nextstep.di.factory.BeanFactory;
-import nextstep.di.factory.ClasspathBeanScanner;
-import nextstep.di.factory.ConfigurationBeanScanner;
 import nextstep.mvc.HandlerMapping;
 import nextstep.web.annotation.RequestMapping;
 import nextstep.web.annotation.RequestMethod;
